Enable arcade-powered-sourced-build for local and CI builds
This enables 'source-build', which makes it easier to build the entire
shipping .NET SDK from source.

This is the first and second step of arcade-powered-source-build:
https://github.com/dotnet/source-build/blob/master/Documentation/planning/arcade-powered-source-build/README.md

See dotnet/sourcelink#692 for a similar PR, that this is based on.

And try and resolve some nullability errors.

For source-build, sisable warnings so we can build this without
extensive customizations, since we update the TFM ourselves.