Avoid build failures when .NET 462 targeting pack is missing
These two test projects multitarget net462 on Windows, but if the
targeting pack is missing, the build will fail. This can be avoided
by installing the reference assemblies package, which will automatically
provide the right dependencies according to the TFM.

Rather than depending on Microsoft.NETFramework.ReferenceAssemblies.net462,
this adds the metapackage, which will automatically pick the right dependency
if the projects are ever updated (i.e. to net472, which sounds like it would
be a good idea.

Fixes #1252