Switch to storing annotations in a static field
Although it's kind of icky to store instance data on a static
field, it is implemented in a robust manner that should prevent
any surprises, and the details are hidden from authors of CLIs
and authors of subsystems.

A subsystem reference is no longer needed when annotating the
CliSymbol objects in the grammar, which makes construction
of a pipleline much simpler.

The fluent grammar construction API now looks like this:

```csharp
new Option<bool>("--greeting").WithDescription("The greeting")
```

Eventually we will be able to use extension properties:

```csharp
new Option<bool>("--greeting") {
  Description = "The greeting"
}
```

We still support the `IAnnotationProvider` model that allows
advanced CLI authors to lazily or dynamically provide annotations.