Remove type from annotation ID
The purpose of this was to ensure that accessors used the correct
type for the annotation values. This originally applied to the
AnnotationAccessor<T> helper for symbol.With(help.Description, desc)
but also applied to later manually defined accessors.

This strong typing was not perfect, for example it was possible to
(incorrectly) define two annotations with the same ID and different
types, resulting in uninformative cast exceptions.

It also did not apply cleanly to more advanced cases, such as when
the value was a Func<T> or where the annotation value type was based
on the symbol value type. It is also not unreasonable to want to store
more than one type in an annotataion for othe reasons. All these
required removing the typing by using AnnotationId<object?>.

This miakes things a bit cleaner by removing the strong typing from
the annotation ID and instead adding TryGetAnnotation overloads
that take a type and throw an informative exception on mismatch.

This only affects folks defining new annotations. It does not change
anything for authors of CLIs or authors of subsystems, as they should
to use the strongly typed accessors that should be defined along with
every annotation ID.